+++ NEWS.md
@@ -1,4 +1,4 @@
-# medrobust (development version)
+# medrobust 0.4.2 (2026-09-24)
 
 ## Documentation